requirements.txt
numpy==1.19.2
pandas==1.1.3
tensorflow-cpu==2.5.0
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.3.20
scikit-learn==0.23.2
gunicorn==20.1.0
psycopg2==2.8.6
simplejson==3.17.2